Support newrelic_rpm gem instrumentation (#121)
The newrelic_rpm gem expects `:controller` in instrumentation event
payload to point to an actual Rails controller, so it tries calling
`controller_path` on it. I considered putting the Rodauth controller
there, but I found it misleading, since it's the Rodauth application
processing the request, the Rodauth controller is only doing view
rendering. In the end I decided to just define a `controller_path`
method on the Rodauth app.