Additional improvements for fixtures
The Rails model generator also checks whether :fixture_replacement is
nil, which appears to be used by factory_bot_rails, so we do it as well
here. We remove the test for disabling fixtures, because it caused tests
to fail (since it changes global state), and it's not important enough.

We also make sure we generate the fixture with ERB code, instead of
evaluating the ERB on generation. This is more developer-friendly, as
it allows developers to know the password in tests.

We also remove the `BCrypt::Password#to_s`, as it doesn't appear to be
necessary.