Make fixes to Tailwind templates and simplify them
Remove additional markup around the form to center it and provide a title,
as for me the trade-off between simplicity isn't worth it, and centering
is less likely to look good on different layouts.

Emphasize labels more and improve spacing, and fix red input border not
being applied on validation errors, as well as its focus state. We also
match the red color of validation errors with the input border color.
Use cursor pointer for submit buttons on hover, as that's not the default
with Tailwind.

Remove competing utility classes in some cases, as well as "ng-*" Angular
classes that are unrelated to Tailwind. Also remove some unnecessary
utilities that don't change anything, such as some submit buttons being
flexboxes.

In the login form, we make sure the display the readonly login field
correctly, and move the footer links back to the original partial.

In dark mode, instead of setting background and text colors, we rely on
the defaults set in the layout, as that's more likely to fit into the
current design, and it's less code. We also use lighter red color for
validation errors, so that it looks equally glowy as the bright emerald
color.