Fix WebAuthn setup template being incompatible with account verification
When using the webauthn_verify_account feature, the WebAuthn setup form
is being rendered for account verification instead of the usual verify
account form, and the submission should go the verify account route.
However, because we're passing webauthn setup path in the generated
form, this doesn't happen.

The original built-in template doesn't set the form action, so the form
submission goes to whatever the current URL is. The `form_with` helper
doesn't appear to support not passing `:url`, so we pass the current
request path as a workaround.