Add --argon2 option to the install generator
OWASP recommends using Argon2 for password hashing if possible, and
fallback to bcrypt as the second choice. So, I think it's worth making
it easier to start out with Argon2.

The rodauth-rails gem still has a hard dependency on bcrypt, though it's
not loaded if Argon2 is used. I might consider dropping this dependency
in the next major version.

https://cheatsheetseries.owasp.org/cheatsheets/Password_Storage_Cheat_Sheet.html