Additional fixes for --argon2 generator option
Since the assumption is that the install generator will be run on a new
application that doesn't have any accounts yet, we disable loading bcrypt
by default. The developer can easily remove this if there are existing
accounts with password hashed by bcrypt.

We also make the generated fixtures work with argon2. I will probably
try to get the `password_hash` method public to simplify the code.