Move Sequel setup into Rodauth configuration file
This allows loading Sequel extensions that connect to the database on
load (e.g. sql_log_normalizer), while still keeping the app bootable
without a database, so that tasks such as `rails db:create` are runnable
without errors.

One added benefit is that this makes Sequel lazily loaded, which makes
rodauth-rails load faster, reducing the impact on app boot time.

It also further cements the idea that Sequel is just an implementation
detail of Rodauth, and not a first-class citizen with its own
initializer.