Add `table_prefix` shorthand for changing table and column names
In applications that already have a `users` table, or want to have
multiple configurations using separate tables, it's not straightforward
to find all `*_table` and `*_column` config methods that need to be
updated with new names. I used to recommend using metaprogramming to
automatically update table and column names, but that won't work with
subclasses that enable new features.

I tried adding a commented-out list in generated `rodauth_main.rb`, and
it's a very long list. Even with a list, the user needs to remember to
rename tables and columns whenever they add new features, both in the
migration and the configuration, which isn't very ergonomic. Choosing a
different prefix should be more streamlined.

As a first step, we add a `table_prefix` configuration method, which
replaces "account" in all table and column names with the given
identifier. We perform this override only in `post_configure`, so that
it includes any features enabled after calling `table_prefix`, such as
those enabled in an auth subclass. This functionality assumes consistent
naming in external features.

See https://github.com/janko/rodauth-rails/discussions/195