Add table argument to install generator
This completes the story around making it easier to use different table
prefix than "account", by allowing this to be changed already on
installation.

The implementation uses the new `--prefix` option for the migration
generator to forward the specified table. It also ensures the selected
table is used for `table_prefix` in generated configuration.

https://github.com/janko/rodauth-rails/discussions/195