Remove the `table_prefix` configuration method
I gave this some more thought, and I don't think it's a good idea to
provide this shorthand. First and foremost, it's using inexact
metaprogramming to rename table and column names, which could fail if
other configuration methods ending with `*_table` appear. Also, the
current implementation didn't support changing table prefix multiple
times in an inheritance chain. The fact that renaming happens in
`post_configure` means we can't simplify rodauth-model interface to
something like `RodauthMain.model` in the future, because with
autoloading we cannot guarantee `post_configure` was called.

This configuration method would mean we'd be diverging from core
Rodauth, because this functionality would be unrelated to Rails. My goal
is to preserve Rodauth functionality as much as possible, while still
providing the convenience Rails developers are used to. I want to avoid
creating a Rails flavor of Rodauth, beyond changing some defaults for
the sake of convenience.

The table argument for rodauth:install will continue to be supported, it
will just rename individual tables, and show how they can be renamed if
the developer chose the default table names. I'll also extend the
rodauth:migration generator to print out Rodauth configuration that
should be added when a table prefix was specified.