Leverage new `account!` method in `rails_account`
This method was added in Rodauth 2.29, and it returns the account if it
has already been loaded, otherwise it retrieves the account from session
if logged in. We essentially make `rails_account` a wrapper around that
method, which returns the model instance instead of a hash.

This stops clearing the session when the account doesn't exist, which is
technically a backwards incompatible change, but I didn't find practical
use cases for it. My main motivation was for the developer to be able to
get rid of the error by refreshing the page. However, when the error
page is rendered, the session cookie doesn't seem to get updated, so
this never actually worked in practice.

I don't want rodauth-rails methods to differ too much from vanilla
Rodauth behavior, I mostly want to augment original behavior, and
`account!` doesn't clear the session. The gained simplicity of the
`rails_account` method is worth the risk of this being a breaking
change.