Add support for webauthn_autofill feature in generated views
The new webauthn_autofill feature added in Rodauth 2.30 appends the
hidden auth form in the `rodauth.login_form_footer` method, so we need
to use the Rodauth method instead of rendering partials directly. For
consistency we also use `rodauth.login_form_header`. We also need to use
`rodauth.login_field_autocomplete_value` for the autocomplete attribute.

We stop generating views of dependent features, to allow people to
generate views e.g. for webauthn_autofill but not webauthn MFA views, or
otp but not two_factor_base.