Allow referencing custom columns on new accounts
When using `ActiveRecord::Base.instantiate`, accessing a column
attribute that wasn't present in the given hash will raise a
ActiveModel::MissingAttributeError. This is not problematic for
retrieving existing accounts, as the `account` hash will have all the
columns, but it is for new accounts.

This allows things like running model validations in a
`before_create_account` hook instead of Rodauth validations, which is
especially useful when migrating existing authentication to Rodauth.