Make generated `convert_token_id_to_integer?` configuration work for UUIDs
When switching from integer to UUIDs, keeping
`convert_token_id_to_integer? true` will make tokens be silently
rejected, and it won't be obvious at all why that is happening.

So, to reduce surprises, we make the configuration work for string
primary keys as well. On the most recent Rodauth version, this still
should avoid DB queries at boot time when eager loading, which is what
we wanted to achieve with our override.