Fix routes with path format not working in front of http_basic_auth
If a request was made to /my_data.xml with `Accept: */*`, and there was
http basic authentication in front in the Rodauth middleware, the
requested format wouldn't get selected when it came to processing the
controller action. This is because rodauth-rails would infer formats
from the `Accept` header as soon as it instantiated the controller
(which would happen when logging the user in from Authorization header),
preventing Rails from inferring the format from the URL path.

We fix that by only doing this before handling Rodauth endpoints,
leaving requests to Rails endpoints untouched.

Fixes #272