Instrument Rodauth controller and current route
Passing `RodauthApp` for the controller works for logging, but it was
breaking some APM agents like NewRelic, which relied on the constant
name referencing a controller class, so I needed to make the Rodauth
app quack enough like a controller.

We change this by passing the Rodauth controller and current Rodauth
route as the controller and action. This should improve APM
integrations, because it makes it easier to differentiate between
different Rodauth routes. Even though the Rodauth controller is not
actually processing the request, I think this is close enough and prefer
this trade-off.