Add new `rodauth:mailer` generator
Originally, we had the `rodauth:mailer` generator, but it has been
merged to `rodauth:install`, because the total number of emails was
reasonably low. However, the newest version of Rodauth doubles the
number of emails, so it no longer makes sense to generate all email
templates on install.

The new `rodauth:mailer` generator only generates the needed email
templates and mailer methods, and prints Rodauth configuration to
copy over. It also supports generating new email templates when new
features get enabled, in this case mailer methods are printed out as
well, because Thor/Rails don't offer a reliable way to inject public
methods.