Add dedicated configuration for skipping Tilt
Having to pass `render: false` when registering Rodauth configurations can be a
bit awkward when the developer isn't otherwise exposed with Rodauth being a Roda
plugin. Moreover, it can be cumbersome, because it needs to be passed for every
Rodauth configuration.

Having dedicated configuration makes it much more straighforward to skip Tilt,
and this way it doesn't get mixed with things that actually affect
authentication behavior or calling Rodauth methods.